+++ Patchcord/Screens/ContentView.swift
@@ -16,8 +16,6 @@ struct ContentView: View {
         animation: .default)
     private var items: FetchedResults<Item>
 
-    private let ndt7bridge = NDT7Bridge()
-
     var body: some View {
         NavigationView {
             List {
@@ -43,9 +41,6 @@ struct ContentView: View {
                 }
             }
             Text("Select an item")
-        }
-        .onAppear {
-            ndt7bridge.startTest()
         }
     }
