+++ Patchcord/Screens/History/HistoryState.swift
@@ -1,0 +1,18 @@
+//
+//  HistoryState.swift
+//  Patchcord
+//
+//  Created by Maksym Bilan on 15.06.2022.
+//
+
+import Foundation
+
+struct HistoryState {
+    let isLoading: Bool
+}
+
+extension HistoryState {
+    init() {
+        isLoading = false
+    }
+}
