+++ Patchcord/Screens/History/HistoryStateReducer.swift
@@ -1,0 +1,14 @@
+//
+//  HistoryStateReducer.swift
+//  Patchcord
+//
+//  Created by Maksym Bilan on 15.06.2022.
+//
+
+import Foundation
+
+extension HistoryState {
+    static let reducer: Reducer<Self> = { state, action in
+        return state
+    }
+}
